+++ src/blur_cache.cpp
@@ -33,7 +33,7 @@
 #include <chrono>
 #include <memory>
 
-Q_LOGGING_CATEGORY(BLUR_CACHE, "kwin_effect_better_blur_dx.blur_cache", QtInfoMsg)
+Q_LOGGING_CATEGORY(BLUR_CACHE, "kwin_effect_better_blur_dx_wobbly_api.blur_cache", QtInfoMsg)
 
 
 /**
@@ -221,8 +221,8 @@ std::unique_ptr<BBDX::BlurCache> BBDX::BlurCache::create(BBDX::BlurEffect *effec
 
     blurCache->m_texturePass.shader = KWin::ShaderManager::instance()->generateShaderFromFile(
         KWin::ShaderTrait::MapTexture,
-        BBDX::shaderFilePath(":/effects/better_blur_dx/shaders/vertex.vert"),
-        BBDX::shaderFilePath(":/effects/better_blur_dx/shaders/texture.frag")
+        BBDX::shaderFilePath(":/effects/better_blur_dx_wobbly_api/shaders/vertex.vert"),
+        BBDX::shaderFilePath(":/effects/better_blur_dx_wobbly_api/shaders/texture.frag")
     );
 
     if (!blurCache->m_texturePass.shader) {
@@ -442,6 +442,12 @@ void BBDX::BlurCache::drawCached(const KWin::RenderViewport &viewport, BBDX::Blu
     glDisable(GL_BLEND);
 
     KWin::ShaderManager::instance()->popShader();
+}
+
+void BBDX::BlurCache::finishExternalCache(BBDX::BlurRenderData &renderInfo) const {
+    if (const auto &cacheEntry = renderInfo.cache.get()) {
+        cacheEntry->flushed(m_paintData);
+    }
 }
 
 void BBDX::BlurCache::drawToCache(BBDX::BlurCacheEntry *cache, KWin::GLVertexBuffer *vbo) const {